move current mark datastructures to own module
